requirements.txt
moviepy>=1.0.3
Pillow>=9.0.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.0